+++ hw/xfree86/common/xf86Events.c
@@ -90,8 +90,6 @@
 #include "dpmsproc.h"
 #endif
 
-#include "xf86platformBus.h"
-
 #include "../os-support/linux/systemd-logind.h"
 
 extern void (*xf86OSPMClose) (void);
