+++ hw/xfree86/common/xf86platformBus.h
@@ -37,20 +37,9 @@ struct xf86_platform_device {
 #define XF86_PDEV_PAUSED        0x04
 
 #ifdef XSERVER_PLATFORM_BUS
-extern int xf86_num_platform_devices;
-extern struct xf86_platform_device *xf86_platform_devices;
-
 static inline struct OdevAttributes *
 xf86_platform_device_odev_attributes(struct xf86_platform_device *device)
 {
-    return device->attribs;
-}
-
-static inline struct OdevAttributes *
-xf86_platform_odev_attributes(int index)
-{
-    struct xf86_platform_device *device = &xf86_platform_devices[index];
-
     return device->attribs;
 }
