+++ hw/xfree86/os-support/linux/systemd-logind.c
@@ -38,7 +38,7 @@
 #include "os.h"
 #include "linux.h"
 #include "xf86_priv.h"
-#include "xf86platformBus.h"
+#include "xf86platformBus_priv.h"
 #include "xf86Xinput_priv.h"
 #include "xf86Priv.h"
 #include "globals.h"
